move `zz_generated_crds.go` out of `pkg/apis` to avoid dependencies to controller-manager-library in "pkg/apis" module
