+++ pkg/apis/cert/register.go
@@ -4,6 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-//go:generate sh -c "CONTROLLER_GEN=$CONTROLLER_GEN bash $CONTROLLER_MANAGER_LIB_HACK_DIR/generate-crds"
+//go:generate sh -c "CONTROLLER_GEN=$CONTROLLER_GEN CONTROLLER_MANAGER_LIB_HACK_DIR=$CONTROLLER_MANAGER_LIB_HACK_DIR ../../../hack/generate-crds.sh"
 
 package cert
