+++ pkg/controller/issuer/controller.go
@@ -14,8 +14,8 @@ import (
 	"github.com/gardener/controller-manager-library/pkg/resources"
 	"github.com/gardener/controller-manager-library/pkg/resources/apiextensions"
 
-	"github.com/gardener/cert-management/pkg/apis/cert/crds"
 	api "github.com/gardener/cert-management/pkg/apis/cert/v1alpha1"
+	"github.com/gardener/cert-management/pkg/cert/crds"
 	"github.com/gardener/cert-management/pkg/cert/source"
 	ctrl "github.com/gardener/cert-management/pkg/controller"
 	"github.com/gardener/cert-management/pkg/controller/issuer/core"
